Move TBTC chain interface to `pkg/chain`
